% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AddFactor}
\alias{AddFactor}
\title{Adds a paramter to factor collection}
\usage{
AddFactor(factors = c(), lambda = "qunif", name, min, max)
}
\arguments{
\item{factors}{The current factor collection}

\item{lambda}{The function to apply FUN(p,min,max)}

\item{name}{The name of factor}

\item{min}{The minimun of parameter p}

\item{max}{The maximun of parameter p}
}
\value{
The collection of created factors
}
\description{
Builds up the factor collection.
}
\examples{
\dontrun{
   f<- AddFactor(name="Age",min=20,max=60)
   f<- AddFactor(factors=f, name="Weight",min=50,max=120)}

}
