% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AddFactor0}
\alias{AddFactor0}
\title{AddFactor0}
\usage{
AddFactor0(factors = c(), ...)
}
\arguments{
\item{factors}{The current factor collection}

\item{...}{The variadic parameter list}
}
\value{
The factor collection
}
\description{
Creates or appends the factor collection
}
\examples{
\dontrun{
   f<- AddFactor0(name="Age",min=20,max=60)
   f<- AddFactor0(factors=f, name="Weight",min=50,max=120)}

}
