% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AddResults}
\alias{AddResults}
\title{Concatenate results of multiple runs}
\usage{
AddResults(d)
}
\arguments{
\item{d}{A data frame containing one replication data}
}
\description{
This function stores the output
of the last model execution and it is intended
to be used internally.
}
