% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.Base}
\alias{AoE.Base}
\title{AoE.Base}
\usage{
AoE.Base(m, factors = c(), fun = NULL)
}
\arguments{
\item{m}{The base design matrix}

\item{factors}{A subset of model parameters}

\item{fun}{The function which will be applied to m}
}
\value{
The design matrix
}
\description{
The Design Of Experiments Base function
}

