% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AoE.ColumnCoV}
\alias{AoE.ColumnCoV}
\title{AoE.ColumnCoV}
\usage{
AoE.ColumnCoV(dataset, key)
}
\arguments{
\item{dataset}{A model output dataset}

\item{key}{Column name from output dataset}
}
\value{
A data frame with Coefficient of variations
}
\description{
This function Calculates the relative squared 
deviation (RSD or CoV) for an used provided column name \code{key}
in the parameter \code{dataset}.
}

