% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.CoV}
\alias{AoE.CoV}
\title{AoE.CoV}
\usage{
AoE.CoV(d)
}
\arguments{
\item{d}{The data collection}
}
\value{
The coefficient of variation for data
}
\description{
A simple funcion for calculate the
Coefficient of Variation
}
