% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.FullFactorial}
\alias{AoE.FullFactorial}
\title{AoE.FullFactorial design generator}
\usage{
AoE.FullFactorial(n = 10, factors = c())
}
\arguments{
\item{n}{The number of samples}

\item{factors}{The model's parameters which will be evaluated}
}
\value{
The Full Factorial design matrix for provided parameters
}
\description{
Generate a Full Factorial sampling for evaluating
the parameters of a model.
}
\examples{
\dontrun{
 f<- AddFactor(name="cyclePoint",min=40,max=90)
 f<- AddFactor(factors=f, name="conjugationCost",min=1,max=80)
 d<- AoE.FullFactorial(2,f)}

}
