% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AoE.GetMorrisOutput}
\alias{AoE.GetMorrisOutput}
\title{AoE.GetMorrisOutput}
\usage{
AoE.GetMorrisOutput(obj)
}
\arguments{
\item{obj}{A reference to a morris object instance}
}
\value{
The results of Morris method
}
\description{
Returns a dataframe holding the Morris 
result set
}

