% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{AoE.LatinHypercube}
\alias{AoE.LatinHypercube}
\title{AoE.LatinHypercube}
\usage{
AoE.LatinHypercube(n = 10, factors = c())
}
\arguments{
\item{n}{The number of samples}

\item{factors}{The model's parameters which will be evaluated}
}
\value{
The LHS design matrix for provided parameters
}
\description{
Generate a LHS sample for model parameters
}
\details{
Generate the LHS sampling for evaluating 
the parameters of a model.
}
\examples{
\dontrun{
 f<- AddFactor(name="cyclePoint",min=40,max=90)
 f<- AddFactor(factors=f, name="conjugationCost",min=1,max=80)
 d<- DoE.LatinHypercube(2,f)}

}

