% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.MAE}
\alias{AoE.MAE}
\title{AoE.MAE}
\usage{
AoE.MAE(xs, xe)
}
\arguments{
\item{xs}{The simulated data set}

\item{xe}{The experimental data set}
}
\value{
The MAE value for provided datasets
}
\description{
Calculates the average-error
magnitude (MAE)
}
