% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.Morris}
\alias{AoE.Morris}
\title{AoE.Morris}
\usage{
AoE.Morris(k = c(), p = 5, r = 4)
}
\arguments{
\item{k}{The factors for morris screening.}

\item{p}{The number of levels for the model's factors.}

\item{r}{Repetitions. The number of random sampling points of Morris Method.}
}
\description{
This is a wrapper for performing Morris's  screening
method on repast models. We rely on morris method from sensitivity
package.
}
\references{
Gilles Pujol, Bertrand Iooss, Alexandre Janon with contributions from Sebastien Da Veiga, Jana Fruth,
Laurent Gilquin, Joseph Guillaume, Loic Le Gratiet, Paul Lemaitre, Bernardo Ramos and Taieb Touati (2015).
sensitivity: Sensitivity Analysis. R package version 1.11.1.
https://CRAN.R-project.org/package=sensitivity
}
