% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.NRMSD}
\alias{AoE.NRMSD}
\title{AoE.NRMSD}
\usage{
AoE.NRMSD(xs, xe)
}
\arguments{
\item{xs}{The simulated data set}

\item{xe}{The experimental data set}
}
\value{
The NRRMSD value for provided datasets
}
\description{
A simple Normalized Root-Mean-Square
Deviation calculation using max and min values.
NRMSD = RMSD(x) / (max(x) - min(x))
}

