% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.RandomSampling}
\alias{AoE.RandomSampling}
\title{AoE.RandomSampling experiment desing generator}
\usage{
AoE.RandomSampling(n = 10, factors = c())
}
\arguments{
\item{n}{The number of samples}

\item{factors}{The model's parameters which will be evaluated}
}
\value{
The random sampling design matrix
}
\description{
Generate a Simple Random Sampling experiment design
matrix.
}
\examples{
\dontrun{
 f<- AddFactor(name="cyclePoint",min=40,max=90)
 f<- AddFactor(factors=f, name="conjugationCost",min=1,max=80)
 d<- AoE.RandomSampling(2,f)}

}
