% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.RMSD}
\alias{AoE.RMSD}
\title{AoE.RMSD}
\usage{
AoE.RMSD(xs, xe)
}
\arguments{
\item{xs}{The simulated data set}

\item{xe}{The experimental data set}
}
\value{
The RMSD value for provided datasets
}
\description{
A simple Root-Mean-Square Deviation
calculation.
}
