% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.Sobol}
\alias{AoE.Sobol}
\title{AoE.Sobol}
\usage{
AoE.Sobol(n = 100, factors = c(), o = 2, nb = 100,
  fun.doe = AoE.LatinHypercube, fun.sobol = sobolmartinez)
}
\arguments{
\item{n}{The number of samples}

\item{factors}{The model's parameters which will be evaluated}

\item{o}{Maximum order in the ANOVA decomposition}

\item{nb}{Number of bootstrap replicates}

\item{fun.doe}{The sampling function to be used for sobol method}

\item{fun.sobol}{The sobol implementation}
}
\description{
This is a wrapper for performing Global Sensitivity
Analysis using the Sobol Method provided by sensitivity
package.
}
\details{
This function is not intended to be used directly from
user programs.
}
\references{
Gilles Pujol, Bertrand Iooss, Alexandre Janon with contributions from Sebastien Da Veiga, Jana Fruth,
Laurent Gilquin, Joseph Guillaume, Loic Le Gratiet, Paul Lemaitre, Bernardo Ramos and Taieb Touati (2015).
sensitivity: Sensitivity Analysis. R package version 1.11.1.
https://CRAN.R-project.org/package=sensitivity
}

