% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-aoe.R
\name{AoE.Stability}
\alias{AoE.Stability}
\title{AoE.Stability}
\usage{
AoE.Stability(dataset, keys = c())
}
\arguments{
\item{dataset}{A model output dataset}

\item{keys}{A list of column names}
}
\value{
A data frame with Coefficient of variations
}
\description{
This function verifies the stability
of CoV for all columns given by parameter \code{keys}
or all dataset columns if keys is empty.
}

