% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{ApplyFactorRange}
\alias{ApplyFactorRange}
\title{Corrects the LHS design matrix}
\usage{
ApplyFactorRange(design, factors)
}
\arguments{
\item{design}{The LHS design matrix}

\item{factors}{THe collection of factors}
}
\value{
The corrected design matrix
}
\description{
Correct the LHS sampling matrix for a
specific range applying the lambda function. The default
value of 'lambda' is 'qunif'.
}
