% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Calibration.GetMemberKeys}
\alias{Calibration.GetMemberKeys}
\title{Calibration.GetMemberKeys}
\usage{
Calibration.GetMemberKeys(obj)
}
\arguments{
\item{obj}{An instance of the object returned by \code{Easy} methods}
}
\value{
The collection of keys
}
\description{
Gets the list of keys (the factor names)
}

