% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{Calibration.GetMemberList}
\alias{Calibration.GetMemberList}
\title{Calibration.GetMemberList}
\usage{
Calibration.GetMemberList(obj, key, name)
}
\arguments{
\item{obj}{An instance of the object returned by \code{Easy} methods}

\item{key}{The key value}

\item{name}{The column name}
}
\value{
The member list
}
\description{
Gets the member list value
}

