% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{check.integration}
\alias{check.integration}
\title{check.integration}
\usage{
check.integration(modelpath)
}
\arguments{
\item{modelpath}{The path where model is installed}
}
\value{
TRUE if the integration code is correctelly deployed
}
\description{
Check if the integration jar library is correctelly installed 
in the model lib directory
}

