% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{check.scenario}
\alias{check.scenario}
\title{check.scenario}
\usage{
check.scenario(modelpath)
}
\arguments{
\item{modelpath}{The path where model is installed}
}
\value{
TRUE if scenario is properly configured
}
\description{
Check if the scenario.xml is configured with the rrepast
itegration code
}

