% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{col.sum}
\alias{col.sum}
\title{col.sum}
\usage{
col.sum(d, skip = c())
}
\arguments{
\item{d}{The data frame}

\item{skip}{The columns which should not be included in the sum}
}
\value{
The original data frame with a new column (sum) holding the sum
}
\description{
Sum all columns but one (pset) of a data frame
}

