% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{config.check}
\alias{config.check}
\title{config.check}
\usage{
config.check(modelpath)
}
\arguments{
\item{modelpath}{The path where model is installed}
}
\value{
TRUE when all requisites are met
}
\description{
Verify if the installed model is correctelly configurated.
}

