% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{config.copylib}
\alias{config.copylib}
\title{config.copylib}
\usage{
config.copylib(modelpath, uninstall = FALSE)
}
\arguments{
\item{modelpath}{The path where model is installed}

\item{uninstall}{If TRUE uninstall integration jar}
}
\description{
Install or uninstall the integration jar file. This function 
manages the installation process of required jars to the model lib dir.
}

