% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{config.scenario}
\alias{config.scenario}
\title{config.scenario}
\usage{
config.scenario(modelpath, uninstall = FALSE)
}
\arguments{
\item{modelpath}{The path where model is installed}

\item{uninstall}{If TRUE restore original scenario.xml file}
}
\value{
A logical TRUE if the model's scenario file has been modified
}
\description{
Add the integration library to the model's configuration
}
