% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{createOutputDir}
\alias{createOutputDir}
\title{Create output directory}
\usage{
createOutputDir()
}
\description{
A simple function to make a directory to save the 
model's data.
}
\details{
Create the, if required, the directory to save the 
output data generate by the model. It is intended for internal 
use.
}

