% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{df2matrix}
\alias{df2matrix}
\title{df2matrix}
\usage{
df2matrix(d, n = c())
}
\arguments{
\item{d}{The data frame}

\item{n}{The column names to be converted. Null for all data frame columns}
}
\value{
The data frame converted to a matrix
}
\description{
This function converts data frames to matrix data type.
}

