% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{dffilterby}
\alias{dffilterby}
\title{dffilterby}
\usage{
dffilterby(d, key, values = c())
}
\arguments{
\item{d}{The data frame holding data to be filtered}

\item{key}{The column name for selection valuas}

\item{values}{The collection of values used to filter the data set}
}
\value{
The filtered data set
}
\description{
Selects a subset of a data frame, filtering by
column values.
}

