% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{dfround}
\alias{dfround}
\title{dfround}
\usage{
dfround(d, p)
}
\arguments{
\item{d}{The data frame}

\item{p}{The number of decimal digits to be keept}
}
\value{
A data frame with rounded columns
}
\description{
Round all numeric columns of a data frame
}

