% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.getChart}
\alias{Easy.getChart}
\title{Easy.getChart}
\usage{
Easy.getChart(obj, key)
}
\arguments{
\item{obj}{A reference to the output of Easy.Stability}

\item{key}{The param name}
}
\value{
The plot instance
}
\description{
Returns the chart instance
}
