% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.getPlot}
\alias{Easy.getPlot}
\title{Easy.getPlot}
\usage{
Easy.getPlot(obj, c, key)
}
\arguments{
\item{obj}{A reference to the output of an "Easy" API method}

\item{c}{The output name}

\item{key}{The param name}
}
\value{
The plot instance
}
\description{
Returns the chart instance
}
