% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.Morris}
\alias{Easy.Morris}
\title{Easy API for Morris's screening method}
\usage{
Easy.Morris(m.dir, m.ds, m.time = 300, parameters, mo.p, mo.r, exp.r, FUN,
  default = NULL)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{parameters}{The factors for morris screening.}

\item{mo.p}{The number of levels for the model's factors.}

\item{mo.r}{Repetitions. The number of random sampling points of Morris Method.}

\item{exp.r}{The number of experiment replications}

\item{FUN}{The calibration function.}

\item{default}{The alternative values for parameters which should be kept fixed}
}
\value{
A list with holding experimnt, object and charts
}
\description{
This function wraps all calls to perform Morris method.
}

