% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.Run}
\alias{Easy.Run}
\title{Easy API for running a model}
\usage{
Easy.Run(m.dir, m.ds, m.time = 300, r = 1, default = NULL)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{r}{The number of replications}

\item{default}{The alternative values for the default model parameters}
}
\description{
This function provides a simple wrapper for performing a single 
or replicated model execution with a single set of parameters.
}
