% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.RunExperiment}
\alias{Easy.RunExperiment}
\title{Easy API for Runnning Experiments}
\usage{
Easy.RunExperiment(m.dir, m.ds, m.time = 300, r = 1, design, FUN,
  default = NULL)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{r}{The number of replications}

\item{design}{The design matrix holding parameter sampling}

\item{FUN}{The objective or cost function. A function defined over the model output.}

\item{default}{The alternative values for parameters which should be kept fixed}
}
\value{
The experiment results
}
\description{
This function provides a simple wrapper for performing experimental
setups using a design matrix
}
