% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Easy.Setup}
\alias{Easy.Setup}
\title{Easy.Setup}
\usage{
Easy.Setup(exp.dir, mdl.dir)
}
\arguments{
\item{exp.dir}{The base directory for storing logs and 
simulation datasets, if omited the mdl.dir is used as base diretory 
instead}

\item{mdl.dir}{The installation directory of some repast model}
}
\description{
This function configures automatically the simulation 
execution environment including the modification of scenario file 
and the creation of output directory.
}

