% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Easy.ShowModelParameters}
\alias{Easy.ShowModelParameters}
\title{Easy.ShowModelParameters}
\usage{
Easy.ShowModelParameters(v)
}
\arguments{
\item{v}{The installation directory of some repast model}
}
\value{
The model parameters
}
\description{
Returns the list current model parameters
}

