% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Easy.Sobol}
\alias{Easy.Sobol}
\title{Easy API for Sobol's SA method}
\usage{
Easy.Sobol(m.dir, m.ds, m.time = 300, parameters, exp.n = 100, exp.r = 1,
  FUN)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{parameters}{-- The input factors}

\item{exp.n}{-- The experiment sample size}

\item{exp.r}{-- The number of experiment replications}

\item{FUN}{-- THe calibration function.}
}
\value{
A list with holding experimnt, object and charts
}
\description{
This functions wraps all required calls to perform 
Sobol method for global sensitivity analysis.
}

