% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Easy.Stability}
\alias{Easy.Stability}
\title{Easy API for output stability}
\usage{
Easy.Stability(m.dir, m.ds, m.time = 300, parameters, samples = 1,
  tries = 100, vars = c(), FUN)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{parameters}{The factors or model's parameter list}

\item{samples}{The number of factor samples.}

\item{tries}{The number of experiment replications}

\item{vars}{The model's output variables for compute CoV}

\item{FUN}{The calibration function.}
}
\value{
A list with holding experimnt, object and charts
}
\description{
This functions run model several times in order to determine 
how many experiment replications are required for model's output being stable
(i.e. the convergence of standard deviation)
}

