% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.Finish}
\alias{Engine.Finish}
\title{Performs a cleanup on engine instance.}
\usage{
Engine.Finish(e)
}
\arguments{
\item{e}{-- An engine object instance}
}
\description{
Finalize and destroy repast controller data.
}

