% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.getId}
\alias{Engine.getId}
\title{Returns the model id}
\usage{
Engine.getId(e)
}
\arguments{
\item{e}{An engine object instance}
}
\description{
This function provides a wrapper to the method getId()
from repast context. The id is basically a String with the currently
instantiated model name.
}
