% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.GetModelOutput}
\alias{Engine.GetModelOutput}
\title{Gets the model output data as a CSV String array}
\usage{
Engine.GetModelOutput(e)
}
\arguments{
\item{e}{-- An engine object instance}
}
\value{
An array of strings containing the model's output
}
\description{
Call the engine method GetModelOutput to drain model output
data.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Engine(d)
   csv<- RepastModel.GetOutput(m)}
}

