% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameter}
\alias{Engine.getParameter}
\title{Get the value of model parameter s as java.lang.Object}
\usage{
Engine.getParameter(e, k)
}
\arguments{
\item{e}{-- An engine object instance}

\item{k}{-- The parameter name}
}
\value{
The parameter value
}
\description{
Get the value of model parameter s as java.lang.Object
}

