% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameterAsNumber}
\alias{Engine.getParameterAsNumber}
\title{Get the value of model parameter s as java.lang.Number}
\usage{
Engine.getParameterAsNumber(e, k)
}
\arguments{
\item{e}{-- An engine object instance}

\item{k}{-- The parameter name}
}
\description{
Get the value of model parameter s as java.lang.Number
}

