% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameterNames}
\alias{Engine.getParameterNames}
\title{Parameter names}
\usage{
Engine.getParameterNames(e)
}
\arguments{
\item{e}{-- An engine object instance}
}
\description{
Returns the names of all declared model's parameters in
the parameter.xml file in the scenario directory.
}

