% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameterType}
\alias{Engine.getParameterType}
\title{Get the type of a model parameter}
\usage{
Engine.getParameterType(e, k)
}
\arguments{
\item{e}{-- An engine object instance}

\item{k}{-- The parameter name}
}
\value{
The paramter type
}
\description{
Get the type of a model parameter
}

