% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine}
\alias{Engine}
\title{Creates an instance of Engine}
\usage{
Engine()
}
\value{
An onject instance of Engine class
}
\description{
This function creates an instance of Repast model wrapper 
class. Before invoking the function Engine, make sure that 
environment was correctly initialized.
}

