% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.resetModelOutput}
\alias{Engine.resetModelOutput}
\title{Engine.resetModelOutput}
\usage{
Engine.resetModelOutput(e)
}
\arguments{
\item{e}{An engine object instance}
}
\description{
Resets the the model output holder
}
