% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.SetAggregateDataSet}
\alias{Engine.SetAggregateDataSet}
\title{Sets the model's dataset}
\usage{
Engine.SetAggregateDataSet(e, k)
}
\arguments{
\item{e}{-- An engine object instance}

\item{k}{-- The repast model's data set name}
}
\description{
Configure a dataset with the desired output values 
to be "drained" by the function Engine.GetModelOutput.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- RepastModel(d)
   setAggregateDataSet(m,"dataset-name")}
}

