% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{enginestats.calls}
\alias{enginestats.calls}
\title{enginestats.calls}
\usage{
enginestats.calls(increment = FALSE)
}
\arguments{
\item{increment}{A flag telling to increment and update the counter}
}
\value{
The number of calls to 'Engine.RunModel'
}
\description{
Return the current calls to the 'Engine.RunModel' function
}
