% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{getExperimentDataset}
\alias{getExperimentDataset}
\title{Helper function to get experiment \code{dataset}}
\usage{
getExperimentDataset(e)
}
\arguments{
\item{e}{The experiement object returned by \code{\link{RunExperiment}}}
}
\value{
The reference to \code{dataset} container.
}
\description{
The RunExperiment function returns a list holding
the \code{paramset}, \code{output} and \code{dataset} collection.
The \code{paramset} collection contains the parameters used for
running the experimental setup. The \code{output} has the results
from user provided calibration function. The \code{dataset}
collection has the raw output of 'Repast' aggregated dataset.
}
