% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetFactorLevels}
\alias{GetFactorLevels}
\title{GetFactorLevels}
\usage{
GetFactorLevels(factors, name)
}
\arguments{
\item{factors}{The current factor collection}

\item{name}{The factor name}
}
\value{
Levels
}
\description{
Returns the fator's levels
}
\examples{
\dontrun{
   f<- AddFactor0(name="Age",levels=c(25,30,40,65))
   f<- AddFactor0(factors=f, name="Weight",levels=c(60,70,80,90))
   
   GetFactorLevels(factors=f, "Age")}

}
