% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetFactorsSize}
\alias{GetFactorsSize}
\title{Get the number of factors}
\usage{
GetFactorsSize(factors)
}
\arguments{
\item{factors}{A collection of factors created with AddFactor}
}
\value{
The number of parameters in factors collection
}
\description{
Returns the total number of factors
}

