% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{getKeyRandom}
\alias{getKeyRandom}
\title{Gets Repast randomSeed name}
\usage{
getKeyRandom()
}
\value{
A string value holding the randomSeed name.
}
\description{
Returns the Repast randomSeed parameter name.
}
