% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetOutput}
\alias{GetOutput}
\title{Gets the output}
\usage{
GetOutput(e)
}
\arguments{
\item{e}{An engine object instance}
}
\value{
Returns a data.frame with output data
}
\description{
Returns the results of a model a data.frame from the last
RUN. Should be used only if model replication is equal to 1,
otherwise GetResults must be used.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Model(d)
   ...
   data<- GetOutput(m)}
   
}

