% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-parallel.R
\name{getpkgcores}
\alias{getpkgcores}
\title{getpkgcores}
\usage{
getpkgcores()
}
\value{
The number of cores
}
\description{
Returns the maximum number of cores to be used in parallel computations
}
