% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-parallel.R
\name{getpkgdefaultcores}
\alias{getpkgdefaultcores}
\title{getpkgdefaultcores}
\usage{
getpkgdefaultcores()
}
\value{
Cores used by R/Repast
}
\description{
Provides the package default parallelism level which is 80\% of total cores available
}
