% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetResultsParameters}
\alias{GetResultsParameters}
\title{Gets the parameters}
\usage{
GetResultsParameters()
}
\value{
A data.frame with parameters of the model.
}
\description{
Returns the current set of paramters used
for the last model run.
}

