% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetSimulationParameters}
\alias{GetSimulationParameters}
\title{Gets the simulation parameters}
\usage{
GetSimulationParameters(e)
}
\arguments{
\item{e}{An engine object instance}
}
\value{
A data frame with simulation parameters
}
\description{
Returns a dataframe with the current set of input 
parameters for the last model run.
}

