% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{GetSimulationParameterType}
\alias{GetSimulationParameterType}
\title{GetSimulationParameterType}
\usage{
GetSimulationParameterType(e, k)
}
\arguments{
\item{e}{An instance of 'Engine' object}

\item{k}{The parameter name}
}
\value{
The parameter type as string
}
\description{
Returns the declared parameter type.
}

