% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{GoToWorkDir}
\alias{GoToWorkDir}
\title{GoToWorkDir}
\usage{
GoToWorkDir()
}
\description{
Changes the current work directory saving the previous one 
which is used in \code{\link{GoToPreviousDir}}. This function is called 
by \code{\link{Easy.Setup}}
}
