% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{hybrid.distance}
\alias{hybrid.distance}
\title{hybrid.distance}
\usage{
hybrid.distance(value, reference, FUN = AoE.NRMSD)
}
\arguments{
\item{value}{The value which will be compared against the reference}

\item{reference}{The reference value. It should be a list holding the value, the range of values.}

\item{FUN}{The distance function. The default is the NRMSD}
}
\value{
The distance metric
}
\description{
Calculates the distance between some value a reference 
target value. It is an hybrid distance because when the value falls
whithin a reference range the distance is 0, otherwise the distance 
between the value and the reference value is calculated using the user
provided distance function.
}

