% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{hybrid.value}
\alias{hybrid.value}
\title{hybrid.value}
\usage{
hybrid.value(value, distance)
}
\arguments{
\item{value}{The reference value}

\item{distance}{The distance interval.}
}
\value{
The list holding the value and the interval 'min --- value --- max'
}
\description{
A simple helper function for generating the input list for 
the function 'hybrid.distance'. This list must hold the value and a range
centered over the value.
}

