% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{jarfile}
\alias{jarfile}
\title{jarfile}
\usage{
jarfile(fjar)
}
\arguments{
\item{fjar}{The name of jar file}
}
\value{
The full path to jar file
}
\description{
The jarfile returns the full path to some jar file 
available inside rrpast package
}

