% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{jvm.enablejmx}
\alias{jvm.enablejmx}
\title{jvm.enablejmx}
\usage{
jvm.enablejmx()
}
\description{
Enable jmx for the current R/rJava session
}
\details{
Configures the JMX subsystem for the current session of R/rJava. 
This function must be called before any other function which initializes 
r/Java such as \code{\link{Easy.Setup}} or \code{\link{Model}} otherwise it will have no effect.
}
\examples{
\dontrun{
   jvm.enablejmx()}

}
