% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{jvm.get_parameters}
\alias{jvm.get_parameters}
\title{jvm.get_parameters}
\usage{
jvm.get_parameters()
}
\value{
A string with JVM parameters.
}
\description{
Returns the current java virtual machine parameters
}
