% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{jvm.getruntime}
\alias{jvm.getruntime}
\title{jvm.getruntime}
\usage{
jvm.getruntime()
}
\description{
A wrapper for \code{System.getRuntime()}
}
\details{
A simple wrapper for \code{System.getRuntime()} java method
}
