% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{jvm.init}
\alias{jvm.init}
\title{Init R/JVM environment}
\usage{
jvm.init()
}
\description{
Initialize rJava and repast environment with classpath. This function
is called internally and it is not meant to be used directlly.
}
\details{
The default parameters can be changed as needed calling the
primitive \code{\link{jvm.set_parameters}} befor instantiating the model
engine.
}
\examples{
\dontrun{
     jvm.init()}

}
\references{
[1] rJava: Low-Level R to Java Interface. Low-level interface to Java VM
very much like .C/.Call and friends. Allows creation of objects,
calling methods and accessing fields.
}

