% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{jvm.set_parameters}
\alias{jvm.set_parameters}
\title{jvm.set_parameters}
\usage{
jvm.set_parameters(s)
}
\arguments{
\item{s}{The paramter string to be passed to the underlying JVM}
}
\description{
Configures the jvm parameters
}
\details{
Set the underlying parameters for java virtual machine. The default 
values are "-server -Xms512m -Xmx1024m". These defaults can be changed 
to fit the model requirements.
}
\examples{
\dontrun{
   jvm.set_parameters("-server -Xms512m -Xmx2048m")}
   
}

