% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{jvm.setOut}
\alias{jvm.setOut}
\title{jvm.setOut}
\usage{
jvm.setOut(f)
}
\arguments{
\item{f}{The output file name}
}
\description{
Set the System.out filed to a file
}
\examples{
\dontrun{
   jvm.setOut("/tmp/SysteOut.log")}
   
}

