% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{LatinHypercube}
\alias{LatinHypercube}
\title{Generate a LHS sample for model parameters}
\usage{
LatinHypercube(n = 10, factors = c())
}
\arguments{
\item{n}{-- The number of samples}

\item{factors}{-- The model's parameters which will be evaluated}
}
\value{
The LHS design matrix for provided parameters
}
\description{
Generate the LHS sampling for evaluating 
the parameters of a model.
}

