% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{lcontains}
\alias{lcontains}
\title{lcontains}
\usage{
lcontains(l, n)
}
\arguments{
\item{l}{The list object}

\item{n}{The item name}
}
\value{
Boolean TRUE if name is found on list
}
\description{
Cheks if a list contains a name
}
