% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Load}
\alias{Load}
\title{The Scenario loader}
\usage{
Load(e)
}
\arguments{
\item{e}{An engine object instance}
}
\description{
Loads the model's scenario. This function must be
called before running the model.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Model(d)
   Load(m)}

}

