% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-pb.R
\name{PB.init}
\alias{PB.init}
\title{PB.init}
\usage{
PB.init(psets, replications)
}
\arguments{
\item{psets}{-- The total number of paramter sets being simulated}

\item{replications}{-- The number of replications per simulation round}
}
\description{
Initialize progress bar for model
execution.
}
