% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{PB.isEnabled}
\alias{PB.isEnabled}
\title{PB.isEnabled}
\usage{
PB.isEnabled()
}
\value{
Boolean TRUE if progress bar must be shown
}
\description{
Returns the global value indicating if progress bar 
is enabled.
}

