% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-helper.R
\name{pick.fittest}
\alias{pick.fittest}
\title{pick.fittest}
\usage{
pick.fittest(out, goals = c(), n = 4)
}
\arguments{
\item{out}{The output data set holding the values of goals}

\item{goals}{The column names which must be used as goal}

\item{n}{The number of solutions}
}
\value{
The n rows holding the best results
}
\description{
Choose the best solutions minimizing the objective function
}
