% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Plot.Calibration}
\alias{Plot.Calibration}
\title{Plot of calibration}
\usage{
Plot.Calibration(obj, key, title = NULL)
}
\arguments{
\item{obj}{An instance of calibration Object}

\item{key}{The column name}

\item{title}{Chart title, may be null}
}
\value{
The resulting ggplot2 plot object
}
\description{
Generate plot for parameter sets 
providing best fit
}

