% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-plots.R
\name{Plot.Morris}
\alias{Plot.Morris}
\title{Plot of Morris output}
\usage{
Plot.Morris(obj, type, title = NULL)
}
\arguments{
\item{obj}{An instance of Morris Object \code{\link{AoE.Morris}}}

\item{type}{The chart type (mu*sigma|musigma|mu*mu)}

\item{title}{Chart title, may be null}
}
\value{
The resulting ggplot2 plot object
}
\description{
Generate plot for Morris's screening method
}

