% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Plot.Sobol}
\alias{Plot.Sobol}
\title{Plot of Sobol output}
\usage{
Plot.Sobol(obj, type, title = NULL)
}
\arguments{
\item{obj}{An instance of Sobol Object \code{\link{AoE.Sobol}}}

\item{type}{The chart type}

\item{title}{Chart title, may be null}
}
\value{
The resulting ggplot2 plot object
}
\description{
Generate plot for Sobol's GSA
}

