% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Plot.Stability}
\alias{Plot.Stability}
\title{Plot stability of output}
\usage{
Plot.Stability(obj, title = NULL)
}
\arguments{
\item{obj}{An instance of Morris Object \code{\link{AoE.Morris}}}

\item{title}{Chart title, may be null}
}
\value{
The resulting ggplot2 plot object
}
\description{
Generate plot for visually access the stability of 
coefficient of variation as function of simulation sample size.
}

