% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Results.GetCharts}
\alias{Results.GetCharts}
\title{Results.GetCharts}
\usage{
Results.GetCharts(obj)
}
\arguments{
\item{obj}{An instance of the object returned by \code{Easy} methods}
}
\value{
The charts element inside results
}
\description{
Simplify the access to the charts member
}

