% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Results.GetExperiment}
\alias{Results.GetExperiment}
\title{Results.GetExperiment}
\usage{
Results.GetExperiment(obj)
}
\arguments{
\item{obj}{An instance of the object returned by \code{Easy} methods}
}
\value{
The experiment element inside results
}
\description{
Simplify the access to the experiment member
}

