% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{Results.GetObject}
\alias{Results.GetObject}
\title{Results.GetObject}
\usage{
Results.GetObject(obj)
}
\arguments{
\item{obj}{An instance of the object returned by \code{Easy} methods}
}
\value{
The object element inside results
}
\description{
Simplify the access to the object member
}

