% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Run}
\alias{Run}
\title{Run simulations}
\usage{
Run(e, r = 1)
}
\arguments{
\item{e}{-- An engine object instance}

\item{r}{-- The number of experiment replications}
}
\description{
This function executes the time steps of an 
instantiated model. The number of replications of model 
runs can be specified by the function parameter.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Model(d)
   Load(m)
   Run(m,r=2)}

}

