% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{RunExperiment}
\alias{RunExperiment}
\title{Run an experimental setup}
\usage{
RunExperiment(e, r = 1, design, FUN)
}
\arguments{
\item{e}{-- An engine object instance}

\item{r}{-- The number of experiment replications}

\item{design}{-- The desing matrix holding parameter sampling}

\item{FUN}{-- THe calibration function.}
}
\description{
Run the model multiple times for different parameters
given by design matrix function parameter.
}
\details{
The FUN function must return zero for perfect fit and values 
greater than zero otherwise.
}

