% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{SaveSimulationData}
\alias{SaveSimulationData}
\title{Saving simulation output}
\usage{
SaveSimulationData(as = "csv")
}
\arguments{
\item{as}{The desired output type, must be csv or xls}
}
\description{
Saves the simulation results of last call to Run(e)
function.
}

