% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{SequenceItem}
\alias{SequenceItem}
\title{SequenceItem}
\usage{
SequenceItem(v, min, max)
}
\arguments{
\item{v}{A column of n x k design matrix}

\item{min}{The lower boundary of range}

\item{max}{The uper boundary of range}
}
\value{
A sequence between min and max value
}
\description{
Generate a sequence from min to max using an increment
based on the number of of elements in v
}

