% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{setKeyRandom}
\alias{setKeyRandom}
\title{Sets Repast randomSeed name}
\usage{
setKeyRandom(k)
}
\arguments{
\item{k}{The string with an alternative name for randomSeed}
}
\description{
Configures a non-default value for Repast randomSeed
parameter name.
}

