% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{setOutputDir}
\alias{setOutputDir}
\title{Sets output directory}
\usage{
setOutputDir(s)
}
\arguments{
\item{s}{The full path for output directory}
}
\description{
Configure the desired directoy to save model
output data.
}

