% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-parallel.R
\name{setpkgcores}
\alias{setpkgcores}
\title{setpkgcores}
\usage{
setpkgcores(v)
}
\arguments{
\item{v}{The number of cores}
}
\description{
Configures the maximum number of cores to be used in parallel computations
}
