% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{SetResultsParameters}
\alias{SetResultsParameters}
\title{Sets the parameters}
\usage{
SetResultsParameters(d)
}
\arguments{
\item{d}{A data.frame with parameter values}
}
\description{
Save the current set of paramters used 
for the last model run.
}

