% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{SetSimulationParameter}
\alias{SetSimulationParameter}
\title{SetSimulationParameter}
\usage{
SetSimulationParameter(e, key, value)
}
\arguments{
\item{e}{An engine object instance}

\item{key}{The paramter name}

\item{value}{The parameter value}
}
\description{
Modify model's default parameter collection
}
