% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{SetSimulationParameters}
\alias{SetSimulationParameters}
\title{Set parameters for running model}
\usage{
SetSimulationParameters(e, p)
}
\arguments{
\item{e}{An engine object instance}

\item{p}{A data frame with simulation parameters}
}
\description{
Modify the repast model parameters with
values provided in parameter 'p' which is a data frame
with just one row.
}
