% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{ShowModelPaths}
\alias{ShowModelPaths}
\title{Prints the paths}
\usage{
ShowModelPaths()
}
\description{
Shows the directories currently used to load model scenario and lib. 
This function is informational only and can be used to check whether 
model data is being loaded properly from correct locations.
}
\examples{
\dontrun{
   ShowModelPaths()}
}

