% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-easyapi.R
\name{simple.fitting}
\alias{simple.fitting}
\title{simple.fitting}
\usage{
simple.fitting(m.dir, m.ds, m.time = 300, parameters, samples = 100,
  tries = 1, design = "lhs", smax = 4, objective)
}
\arguments{
\item{m.dir}{The installation directory of some repast model}

\item{m.ds}{The name of any model aggregate dataset}

\item{m.time}{The total simulated time}

\item{parameters}{The input factors}

\item{samples}{The experiment sample size}

\item{tries}{The number of experiment replications}

\item{design}{The sampling scheme ["lhs"|"mcs"|"ffs"]}

\item{smax}{The number of solutions to be generated}

\item{objective}{The calibration function.}
}
\description{
Simple calibration method. Run an experimental setup and select the 
the best results minimizing the calibration function
}
