% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{UpdateDefaultParameters}
\alias{UpdateDefaultParameters}
\title{UpdateDefaultParameters}
\usage{
UpdateDefaultParameters(e, p)
}
\arguments{
\item{e}{An engine object instance}

\item{p}{The collection of model fixed paramters to change}
}
\description{
Modify the value of the default parameters
which should be kept fixed
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Model(d)
   Load(m)

   p<- c(name1=value1, name2=2)
   UpdateDefaultParameters(m,p)}

}

