% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-parallel.R
\name{WrapperRun}
\alias{WrapperRun}
\title{WrapperRun}
\usage{
WrapperRun(modeldir, datasource, maxtime, r = 1, seed = c(),
  design = NULL, default = NULL, multi = TRUE)
}
\arguments{
\item{modeldir}{The installation directory of some repast model}

\item{datasource}{The name of any model aggregate dataset}

\item{maxtime}{The total simulated time}

\item{r}{The number of experiment replications}

\item{seed}{The random seed collection}

\item{design}{The desing matrix holding parameter sampling}

\item{default}{The alternative values for parameters which should be kept fixed}

\item{multi}{allows forcing single core execution, default is using multi-core}
}
\value{
The model output dataset
}
\description{
Wrapper for the Run and ParallelRun functions
}

