% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-parallel.R
\name{WrapperRunExperiment}
\alias{WrapperRunExperiment}
\title{WrapperRunExperiment}
\usage{
WrapperRunExperiment(modeldir, datasource, maxtime, r = 1, design, FUN,
  default = NULL)
}
\arguments{
\item{modeldir}{The installation directory of some repast model}

\item{datasource}{The name of any model aggregate dataset}

\item{maxtime}{The total simulated time}

\item{r}{The number of experiment replications}

\item{design}{The desing matrix holding parameter sampling}

\item{FUN}{The objective function.}

\item{default}{The alternative values for parameters which should be kept fixed}
}
\value{
The model output dataset
}
\description{
Wrapper for the RunExperiment and ParallelRunExperiment 
functions
}
