% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Easy.Setup}
\alias{Easy.Setup}
\title{Easy.Setup}
\usage{
Easy.Setup(model, deployment = c())
}
\arguments{
\item{model}{The base directory where Repast model is installed.}

\item{deployment}{The directory to save the output and logs.}
}
\description{
This function configures the deployment directory 
where logs and output dataset will be generated.  By default 
the deployment directory will be created under the model 
installation directory. The output generated by the Repast model 
will be redirected to the SystemOut.log file.
}
\details{
If the deployment directory is empty the installation 
directory given by the parameter \code{model} is used instead as 
the base directory. The deployment directory is \code{/rrepast-deployment/}.
}

