% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.endAt}
\alias{Engine.endAt}
\title{Engine.endAt}
\usage{
Engine.endAt(e, v)
}
\arguments{
\item{e}{An engine object instance}

\item{v}{The number of Repast time ticks}
}
\description{
Configure the maximun simulated time for 
the current model run
}

