% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.Finish}
\alias{Engine.Finish}
\title{Engine.Finish}
\usage{
Engine.Finish(e)
}
\arguments{
\item{e}{An engine object instance}
}
\description{
Performs a cleanup on a engine instance.Finalize
and destroy repast controller data.
}

