% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.GetModelOutput}
\alias{Engine.GetModelOutput}
\title{Engine.GetModelOutput}
\usage{
Engine.GetModelOutput(e)
}
\arguments{
\item{e}{An engine object instance}
}
\value{
An array of strings containing the model's output
}
\description{
Gets the model output data as a CSV String array.
Calls the engine method GetModelOutput to drain model output
data.
}
\examples{
\dontrun{
   d<- "c:/usr/models/your-model-directory"
   m<- Model(d)
   csv<- Engine.GetModelOutput(m)}
}
