% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameter}
\alias{Engine.getParameter}
\title{Engine.getParameter}
\usage{
Engine.getParameter(e, k)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}
}
\value{
The parameter value
}
\description{
The function gets the value of model 
parameter \code{k} as java.lang.Object
}

