% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.getParameterAsDouble}
\alias{Engine.getParameterAsDouble}
\title{Engine.getParameterAsDouble}
\usage{
Engine.getParameterAsDouble(e, k)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}
}
\value{
The parmeter value as double
}
\description{
Get the value of model parameter \code{k} as \code{java.lang.Double}
}

