% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.getParameterAsNumber}
\alias{Engine.getParameterAsNumber}
\title{Engine.getParameterAsNumber}
\usage{
Engine.getParameterAsNumber(e, k)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}
}
\value{
The parmeter value as number
}
\description{
Get the value of model parameter
\code{k} as \code{java.lang.Number}
}
