% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.getParameterAsString}
\alias{Engine.getParameterAsString}
\title{Engine.getParameterAsString}
\usage{
Engine.getParameterAsString(e, k)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}
}
\value{
The parameter value as string
}
\description{
Get the value of model parameter \code{k}
as \code{java.lang.String}
}
