% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.getParameterNames}
\alias{Engine.getParameterNames}
\title{Engine.getParameterNames}
\usage{
Engine.getParameterNames(e)
}
\arguments{
\item{e}{An engine object instance}
}
\value{
A collection of parameter names
}
\description{
Get the parameter names
}
\details{
Returns the names of all declared model's parameters in
the parameter.xml file in the scenario directory.
}

