% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine.getParameterType}
\alias{Engine.getParameterType}
\title{Engine.getParameterType}
\usage{
Engine.getParameterType(e, k)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}
}
\value{
The parameter type string
}
\description{
Returns the declared type of a Repast
model parameter
}

