% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.LoadModel}
\alias{Engine.LoadModel}
\title{Engine.LoadModel}
\usage{
Engine.LoadModel(e, f)
}
\arguments{
\item{e}{An engine object instance}

\item{f}{The full path of scenario directory}
}
\description{
Loads the model's scenario files
}
\details{
This function loads the scenario of a Repast Model and 
initialize de model.
}

