% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{Engine}
\alias{Engine}
\title{Engine}
\usage{
Engine()
}
\value{
An onject instance of Engine class
}
\description{
Creates an instance of Engine
}
\details{
This function creates an instance of Repast model wrapper
class. Before invoking the function Engine, make sure that
environment was correctly initialized.
}

