% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Engine.setParameter}
\alias{Engine.setParameter}
\title{Engine.setParameter}
\usage{
Engine.setParameter(e, k, v)
}
\arguments{
\item{e}{An engine object instance}

\item{k}{The parameter name}

\item{v}{The parameter value}
}
\description{
Set the value of model parameter
}

