% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{parallelize}
\alias{parallelize}
\title{parallelize}
\usage{
parallelize(v = NULL)
}
\value{
Boolean with current state
}
\description{
Tells R/Repast to use multicore. Default 
is using just one core.
}

