% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRepast.R
\name{Run}
\alias{Run}
\title{Run simulations}
\usage{
Run(e, r = 1, seed = c())
}
\arguments{
\item{e}{An engine object instance}

\item{r}{The number of experiment replications}

\item{seed}{The random seed collection}
}
\value{
The model output dataset
}
\description{
This function executes the time steps of an
instantiated model. The number of replications of model
runs can be specified by the function parameter. The seed
parameter may be omitted and will be generated internally.
If provided, the seed collection, must contain the same
number of \code{r} parameter.
}
\examples{
\dontrun{
   d<- "C:/usr/models/your-model-directory"
   m<- Model(d)
   Load(m)
   Run(m,r=2) # or Run(m,r=2,seed=c(1,2))}

}
