% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrepast-engine.R
\name{ShowModelPaths}
\alias{ShowModelPaths}
\title{ShowModelPaths}
\usage{
ShowModelPaths()
}
\description{
Prints the paths. Shows the directories
currently used to load model scenario and lib. The output of
this function is informational only and can be used to check
whether model data is being loaded properly from
correct locations.
}
\examples{
\dontrun{
   ShowModelPaths()}

}
